oh god what didnt i do in this commit
ok so i added proper email deletion, got a compiler to turn my html and css into inline to send it with the email. \nI also made static serving of success and error pagpes when deleting entries\nlastly i edited the listing controller to redirect to pages if an error happens. \n i lied i also changed the schedulers to happen at 6pm utc which is midnight for us